requirements.txt
fastapi
uvicorn
python-dotenv
langchain-google-genai
langchain-pinecone
langchain-community
pypdf
python-multipart
mangum
pydantic
pinecone-client